--- MyList.rb
@@ -1,21 +1,0 @@
-require_relative 'My_Enumerable'
-
-class MyList
-  include MyEnumerable
-
-  def initialize(*list)
-    @list = list
-  end
-
-  def each
-    @list.each { |elem| yield(elem) }
-  end
-end
-
-list = MyList.new(1, 2, 3, 4)
-
-puts list.all? {|e| e < 5 } # true
-puts list.all? {|e| e > 5 } # false
-puts list.any? {|e| e == 2} # true
-puts list.any? {|e| e == 5} # false
-p list.filter {|e| e.even?} # [2, 4]
