+++ my_enumerable.rb
@@ -1,6 +1,6 @@
 module MyEnumerable
   def all?(&block)
-    each { |item| return false if !block.call(item) }
+    each { |item| return false unless block.call(item) }
     true
   end
